% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add two numbers using muladd and hence muladdpow}
\usage{
add(value, a)
}
\arguments{
\item{value}{starting value to transform}

\item{a}{number to add}
}
\value{
final calculated value
}
\description{
Function that adds two numbers together using
the existing \code{muladd()} and hence \code{muladdpow()} functions
}
\examples{
add(1, 2) == 1 + 2

}
