% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mul.R
\name{mul}
\alias{mul}
\title{Multiply two numbers using muladd and hence muladdpow}
\usage{
mul(value, m)
}
\arguments{
\item{value}{starting value to transform}

\item{m}{number to multiply by}
}
\value{
final calculated value
}
\description{
Function that multiplies two numbers together using
the existing \code{muladd()} and hence \code{muladdpow()} functions
}
\examples{
mul(1, 2) == 1 * 2

}
