% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muladd.R
\name{muladd}
\alias{muladd}
\title{Combine multiply and add using muladdpow}
\usage{
muladd(value, m, a)
}
\arguments{
\item{value}{starting value to transform}

\item{m}{number to multiply by}

\item{a}{number to add on}
}
\value{
final calculated value
}
\description{
Function that multiplies two numbers together and adds a third using
the existing \code{muladdpow()} function
}
\examples{
muladd(1, 2, 3) == 1 * 2 + 3

}
