% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muladdpow.R
\name{muladdpow}
\alias{muladdpow}
\title{Combine multiply, add and raise to the power}
\usage{
muladdpow(value, m, a, p)
}
\arguments{
\item{value}{starting value to transform}

\item{m}{number to multiply by}

\item{a}{number to add on}

\item{p}{power to raise total by}
}
\value{
final calculated value
}
\description{
Function that multiplies two numbers together, adds a third and
finally raises to a fourth power all in one function
}
\examples{
muladdpow(1, 2, 3, 4) == (1 * 2 + 3) ^ 4

}
