% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practicalA5-package.R
\docType{package}
\name{practicalA5-package}
\alias{practicalA5-package}
\alias{practicalA5}
\title{Example code for practical A-5}
\description{
\code{practicalA5} is a simple package that demonstrates writing nested
functions - so functions that provide simpler versions of other functions
}
\author{
Richard Reeve <richard.reeve@glasgow.ac.uk>
}
